+++ server/index.js
@@ -4,9 +4,7 @@ var mongoose = require('mongoose');
 var passport = require('passport');
 var secrets = require('./config/secrets');
 var webpack = require('webpack');
-var config = require('../webpack/webpack.config.dev.js');
 var app = express();
-var compiler = webpack(config);
 
 // Find the appropriate database to connect to, default to localhost if not found.
 var connect = function() {
@@ -31,6 +29,8 @@ fs.readdirSync(__dirname + '/models').forEach(function(file) {
 var isDev = process.env.NODE_ENV === 'development';
 
 if (isDev) {
+  var config = require('../webpack/webpack.config.dev-client.js');
+  var compiler = webpack(config);
   app.use(require('webpack-dev-middleware')(compiler, {
     noInfo: true,
     publicPath: config.output.publicPath