+++ webpack/webpack.config.dev-server.js
@@ -1,0 +1,68 @@
+var path = require('path');
+var webpack = require('webpack');
+var assetsPath = path.join(__dirname, '..', 'public', 'assets');
+var ExtractTextPlugin = require("extract-text-webpack-plugin");
+
+var commonLoaders = [
+  {
+    /*
+     * TC39 categorises proposals for babel in 4 stages
+     * Read more http://babeljs.io/docs/usage/experimental/
+     */
+    test: /\.js$|\.jsx$/,
+    loader: 'babel',
+    // Reason why we put this here instead of babelrc
+    // https://github.com/gaearon/react-transform-hmr/issues/5#issuecomment-142313637
+    query: {
+      "presets": ["es2015", "react", "stage-0"]
+    },
+    include: path.join(__dirname, '..', 'app'),
+    exclude: path.join(__dirname, '/node_modules/')
+  },
+  { test: /\.json$/, loader: "json-loader" },
+  {
+    test: /\.(png|jpg|svg)$/,
+    loader: 'url?limit=10000'
+  },
+  { test: /\.html$/, loader: 'html-loader' }
+];
+
+module.exports = {
+    // The configuration for the server-side rendering
+    name: "server-side rendering",
+    context: path.join(__dirname, "..", "app"),
+    entry: {
+      server: "./server"
+    },
+    target: "node",
+    output: {
+      // The output directory as absolute path
+      path: assetsPath,
+      // The filename of the entry chunk as relative path inside the output.path directory
+      filename: "server.js",
+      // The output path from the view of the Javascript
+      publicPath: "/assets/",
+      libraryTarget: "commonjs2"
+    },
+    module: {
+      loaders: commonLoaders.concat([
+           { test: /\.scss$/,
+             loader: ExtractTextPlugin.extract('style-loader', 'css-loader?module&localIdentName=[local]__[hash:base64:5]!postcss-loader!sass?includePaths[]='
+                                               + encodeURIComponent(path.resolve(__dirname, '..', 'app', 'scss')))
+           }
+      ])
+    },
+    resolve: {
+      extensions: ['', '.js', '.jsx', '.scss'],
+      modulesDirectories: [
+        "app", "node_modules"
+      ]
+    },
+    plugins: [
+        // extract inline css from modules into separate files
+        new ExtractTextPlugin("styles/main.css"),
+        new webpack.DefinePlugin({
+          __DEV__: true
+        })
+    ]
+};